Refs #49. refactored the current notebook UI.
